pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bud-model-catalog"
version = "0.1.0"
description = "Multi-source LLM model catalog with cost-accurate pricing"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "respx>=0.21", "ruff>=0.4"]

[tool.hatch.build.targets.wheel]
packages = ["src/bud_model_catalog"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
